requirements.txt
requests-html
selenium
requests
bs4
newspaper3k
language_tool_python
html-text